% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_posthoc.R
\name{ANOVA_posthoc}
\alias{ANOVA_posthoc}
\title{Creer un dataframe qui exclut les outliers d'une base de donnees}
\usage{
ANOVA_posthoc(VD, VI1, VI2, data, nb_fact, alpha, effect)
}
\arguments{
\item{VD}{la VD de la base de donnees}

\item{VI1}{VI1 de la bdd a nettoyer}

\item{VI2}{VI2 de la bdd a nettoyer}

\item{data}{dataframe}

\item{nb_fact}{nombre de facteurs: 0, 1 ou 2}

\item{alpha}{risque alpha}

\item{effect}{VI1, VI2 or VI1*VI2}
}
\value{
resultats du test ANOVA, comme dans SPSS, aux normes APA
}
\description{
Creer un dataframe qui exclut les outliers d'une base de donnees
}
\keyword{APA}
\keyword{homoscedasticite,}
