% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemin_SPSS.R
\name{chemin_SPSS}
\alias{chemin_SPSS}
\title{Chemin SPSS}
\usage{
chemin_SPSS(test)
}
\arguments{
\item{test}{le test pour lequel le chemin doit etre imprime:
Voici tous les tests encodes:
r_pearson,chi_indep,phi,v_cramer,rho_spearman,tau_kendall,biseriel_points,
t_indep,anova_posthoc,CL,t_indep,anova_facto,
chi_ajust,U_MW,KW,
regr_simple,regr_multiple,
normality_test,homoscedasticity_test}
}
\value{
retourne le chemin SPSS, en francais et en anglais
}
\description{
Chemin SPSS
}
\keyword{Chemin}
\keyword{SPSS}
