% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{chi_square}
\alias{chi_square}
\title{chi square test}
\usage{
chi_square(VI1, VI2, nb_fact, pr, alpha)
}
\arguments{
\item{VI1}{premier facteur}

\item{VI2}{deuxieme facteur}

\item{nb_fact}{nombre de facteur}

\item{pr}{probabilites theoriques}

\item{alpha}{risque alpha}
}
\value{
Conclusion de test statistique
}
\description{
chi square test
}
\keyword{APA}
\keyword{chisquare,}
