% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{decision}
\alias{decision}
\title{RH0 ou NRH0?}
\usage{
decision(p, alpha = 0.05)
}
\arguments{
\item{p}{p_valeur}

\item{alpha}{risque alpha}
}
\value{
Conclusion de test statistique
}
\description{
RH0 ou NRH0?
}
\keyword{APA}
\keyword{chisquare,}
