% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KW_test.R
\name{KW_test}
\alias{KW_test}
\title{Test de Kruskal-Wallis}
\usage{
KW_test(VD, VI1, data, alpha)
}
\arguments{
\item{VD}{la VD de la base de donnees}

\item{VI1}{VI1 de la bdd}

\item{data}{dataframe}

\item{alpha}{risque alpha}
}
\value{
resultats du test de Kr-Wallis, comme dans SPSS, aux normes APA
}
\description{
Test de Kruskal-Wallis
}
\keyword{APA}
\keyword{Kruskall-Wallis}
\keyword{test,}
