% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomdetest.R
\name{nomdetest}
\alias{nomdetest}
\title{Creer un dataframe qui exclut les outliers d'une base de donnees}
\usage{
nomdetest(test)
}
\arguments{
\item{test}{test}
}
\value{
resultats du test chi-carre, comme dans SPSS, aux normes APA
}
\description{
Creer un dataframe qui exclut les outliers d'une base de donnees
}
\keyword{APA}
\keyword{chisquare,}
