% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{normality}
\alias{normality}
\title{Creer un dataframe qui exclut les outliers d'une base de donnees}
\usage{
normality(VD, VI1, VI2, nb_fact, pred, alpha)
}
\arguments{
\item{VD}{la VD de la base de donnees}

\item{VI1}{VI1 de la bdd}

\item{VI2}{VI2 de la bdd}

\item{nb_fact}{nombre de facteurs: 0, 1 ou 2}

\item{alpha}{risque alpha}
}
\value{
resultat du test de K-S, comme dans SPSS, aux normes APA
}
\description{
Creer un dataframe qui exclut les outliers d'une base de donnees
}
\keyword{Boxplot}
\keyword{Outliers}
