% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_exclusion.R
\name{outliers_exclusion}
\alias{outliers_exclusion}
\title{Creer un dataframe qui exclut les outliers d'une base de donnees}
\usage{
outliers_exclusion(data, VD, VI1, VI2, pred, nb_fact = 0)
}
\arguments{
\item{data}{base de donnees que l'on veut nettoyer}

\item{VD}{la VD de la base de donnees contenant d'eventuels outliers}

\item{VI1}{VI1 de la bdd a nettoyer}

\item{VI2}{VI2 de la bdd a nettoyer}

\item{pred}{prédicteurs continus}

\item{nb_fact}{nombre de facteurs catégoriels: 0, 1 ou 2}
}
\value{
outliers id, une bdd nettoyee ainsi qu'une conclusion en francais
}
\description{
Creer un dataframe qui exclut les outliers d'une base de donnees
}
\keyword{Boxplot}
\keyword{Outliers}
