% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_biserial.R
\name{r_biserial}
\alias{r_biserial}
\title{correlation biserielle de points}
\usage{
r_biserial(V_dicho, V_scale, alpha = 0.05)
}
\arguments{
\item{V_dicho}{premier facteur}

\item{V_scale}{deuxieme facteur}

\item{alpha}{risque alpha}
}
\value{
Conclusion de test statistique
}
\description{
correlation biserielle de points
}
