% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_pearson.R
\name{r_pearson}
\alias{r_pearson}
\title{correlation de Pearson}
\usage{
r_pearson(V1, V2, alpha = 0.05)
}
\arguments{
\item{V1}{premier facteur}

\item{V2}{deuxieme facteur}

\item{alpha}{risque alpha}
}
\value{
Conclusion de test statistique
}
\description{
correlation de Pearson
}
